web: add base chart class and custom chart for application view
